♻️ templates, pagesを作成、formSchemaなどを別ファイルに分離。
