+++ src/app/(main)/recruits/page.tsx
@@ -1,15 +1,8 @@
 import React from "react";
-import RecruitsIndex from "@/components/organisms/recruits/RecruitsIndex";
+import RecruitTopPage from "@/components/pages/RecruitTopPage";
 
 const AllRecruits = () => {
-	return (
-		<div className="bg-slate-100">
-			<div className="container max-w-[960px] mx-auto p-8 flex flex-col gap-8">
-				<h1 className="text-3xl font-bold">Recruits</h1>
-				<RecruitsIndex />
-			</div>
-		</div>
-	);
+	return <RecruitTopPage />;
 };
 
 export default AllRecruits;
