+++ src/app/api/recruits/[id]/edit/route.ts
@@ -22,14 +22,14 @@ export const GET = async (
 		const recruit = await prisma.recruits.findUnique({
 			where: { id: id },
 			select: {
-				id: true,
+				// id: true,
 				title: true,
 				content: true,
 				isPublished: true,
 			},
 		});
 
-        console.log(recruit);
+        // console.log(recruit);
 
 		return NextResponse.json({ recruit });
 	} catch (error) {
