+++ src/components/pages/RecruitDetailPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const RecruitDetailPage = () => {
+  return (
+    <div>RecruitDetailPage</div>
+  )
+}
+
+export default RecruitDetailPage
